updated Openshift and Kube YAML configuration
