+++ deployments/kube/auto/scripts/cega-mq.sh
@@ -10,9 +10,7 @@ rabbitmq-plugins enable --offline rabbitmq_shovel_management
 
 cp --remove-destination /temp/rabbitmq.config /etc/rabbitmq/rabbitmq.config
 cp --remove-destination /temp/defs.json /etc/rabbitmq/defs.json
-chown rabbitmq:rabbitmq /etc/rabbitmq/rabbitmq.config
 chmod 640 /etc/rabbitmq/rabbitmq.config
-chown rabbitmq:rabbitmq /etc/rabbitmq/defs.json
 chmod 640 /etc/rabbitmq/defs.json
 
 exec rabbitmq-server
