+++ deployments/kube/test/Makefile
@@ -14,12 +14,12 @@ FILE=HG00458.unmapped.ILLUMINA.bwa.CHS.low_coverage.20130415.bam
 DOCKER_PATH=$(MAIN_REPO)/deployments/kube
 PGP_PUB=$(DOCKER_PATH)/auto/config/key.1.pub
 # should be changed with remote ip, this is the default ip for kubernetes
-DEPLOY_IP=192.168.99.100
+DEPLOY_IP=$(shell minikube ip)
 PGP_EMAIL=local-ega@ega.eu
 CEGA_PORT=$(shell kubectl describe svc cega-mq --namespace=testing | grep "NodePort:" | awk '$$2=="amqp" {print substr($$3,1,5)}')
 INBOX_PORT=$(shell kubectl describe svc inbox --namespace=testing | grep "NodePort:" | awk '{print substr($$3,1,5)}')
 #needs the same password that the Lega-MQ conencted to CEGA-MQ
-CEGA_PASWORD=$(shell kubectl exec mq-0 --namespace=testing env | grep "CEGA_CONNECTION=" | awk '{print substr($$1,29,32)}')
+CEGA_PASWORD=$(shell kubectl get secrets cega-connection --namespace=testing -o 'go-template={{index .data "address"}}' | base64 -d | awk '{print substr($$1,13,32)}')
 CEGA_MQ_CONNECTION=amqp://lega:$(CEGA_PASWORD)@$(DEPLOY_IP):$(CEGA_PORT)/lega
 
 ##############################
@@ -34,12 +34,13 @@ $(FILE).c4ga: $(FILE)
 upload: $(FILE).c4ga
 	chmod 400 $(SSH_KEY_PRIV)
 	cd $(<D) && sftp -P $(INBOX_PORT) -i $(SSH_KEY_PRIV) $(USER)@$(DEPLOY_IP) <<< $$"put $(<F)"
+	chmod 700 $(SSH_KEY_PRIV)
 
 $(FILE).c4ga.md5: $(FILE).c4ga
-	printf '%s' $(shell md5 $< | cut -d' ' -f4 || md5sum $< | cut -d' ' -f1) > $@
+	printf '%s' $(shell md5sum $< | cut -d' ' -f1) > $@
 
 $(FILE).md5: $(FILE)
-	printf '%s' $(shell md5 $< | cut -d' ' -f4 || md5sum $< | cut -d' ' -f1) > $@
+	printf '%s' $(shell md5sum $< | cut -d' ' -f1) > $@
 
 submit: $(FILE).c4ga $(FILE).c4ga.md5 $(FILE).md5
 	@echo publish.py --connection amqp://[redacted]@$(lastword $(subst @, ,$(CEGA_MQ_CONNECTION))) $(USER) dir/$(FILE).c4ga --enc ...
