+++ deployments/kube/yml/lega-config/secret.lega.yml
@@ -26,7 +26,7 @@ data:
 apiVersion: v1
 kind: Secret
 metadata:
-  name: cega-keys
+  name: cega-creds
 type: Opaque
 data:
   credentials: <change me>
