+++ deployments/kube/yml/s3/sts.minio.yml
@@ -26,6 +26,7 @@ spec:
               name: s3-keys
               key: secret_key
         image: minio/minio:latest
+        imagePullPolicy: Always
         args:
         - server
         - "--config-dir=/data/config"
