+++ deployments/kube/yml/verify/sts.verify.yml
@@ -23,6 +23,11 @@ spec:
         imagePullPolicy: Always
         command: ["gosu", "lega", "ega-verify"]
         env:
+          - name: POSTGRES_PASSWORD
+            valueFrom:
+              secretKeyRef:
+                name: lega-db-secret
+                key: postgres_password
           - name: LEGA_PASSWORD
             valueFrom:
               secretKeyRef:
